Test: fix SSEHashSetN test containers
had to trim some of them, since they were buggy and I'm really **bored** from this work :p

+++ Source/Tools/UnitTest/Private/Containers/SSEHashSet5.h
@@ -59,13 +59,13 @@ struct ALIGN(16) TSSEHashBucket5 {
         // otherwise use default
         default:
             // first try to align on cache line size
-            for (size_t c = 11; c <= 16; ++c) {
+            for (u32 c = 11; c <= 16; ++c) {
                 if (Meta::IsAligned(CACHELINE_SIZE, sizeof(states_t) + sizeof(storage_t) * c))
                     return c; // succeed, bail out
             }
 
             // then try to align on simd alignment requirement
-            for (size_t c = 11; c <= 16; ++c)
+            for (u32 c = 11; c <= 16; ++c)
                 if (Meta::IsAligned(16, sizeof(states_t) + sizeof(storage_t) * c))
                     return c; // succeed, bail out
